Fixed errors with logfiles. Fixed some cp and mv commands to work correctly. Added a test before creating a symlink to adcprep to avoid an error.

+++ tapis-pylauncher/job_configs/tx_fema_storms/pre_process.sh
@@ -14,9 +14,10 @@ DEBUG=true
 # Read command line inputs
 STORM_NUM=$1
 ADCIRC_RUN_PROC=$2
+LOG_FILE=$(pwd)"/logs/runs/s${STORM_NUM}_pre.log"
 
 log () {
-  echo "$(date) : ${1} - ${2}" >> "logs/runs/s${STORM_NUM}_pre.log"
+  echo "$(date) : ${1} - ${2}" >> $LOG_FILE
 }
 
 if [ "$DEBUG" = true ] ; then
@@ -27,25 +28,25 @@ fi
 log INFO "Starting pre-process for storm ${STORM_NUM}"
 
 # Create Run Dir - Copy Base Inputs TODO: Check preservation of sym links?
-RUN_DIR="runs/s${NUM}"
+RUN_DIR="runs/s${STORM_NUM}"
 cp -r --preserve=links ./mesh $RUN_DIR
 
 # Copy wind and pressure files for strom to run directory 
-cp --preserve=links "storms/TEX_FEMA_RUN$STORM_NUM.pre $RUN_DIR/fort.221"
-cp --preserve=links "storms/TEX_FEMA_RUN$STORM_NUM.WND $RUN_DIR/fort.222"
+cp --preserve=links storms/TEX_FEMA_RUN$STORM_NUM.pre $RUN_DIR/fort.221
+cp --preserve=links storms/TEX_FEMA_RUN$STORM_NUM.WND $RUN_DIR/fort.222
 
 cd $RUN_DIR
 
-# Copy adcprep executables to run directory 
-ln -s ../../adcprep . 
+# Copy adcprep executables to run directory (if needed)
+[ -f adcprep ] || ln -s ../../adcprep .
 
 # Run ADCPREP
 log INFO "Starting ADCPREP first execution."
-printf '%s\n1\nfort.14\n' "$ADCIRC_RUN_PROC" | adcprep >>  "logs/runs/s${STORM_NUM}_pre.log"
+printf '%s\n1\nfort.14\n' "$ADCIRC_RUN_PROC" | ./adcprep >> $LOG_FILE
 log INFO "ADCPREP first execution done."
 
 log INFO "Starting ADCPREP second execution."
-printf '%s\n2\n' "$ADCIRC_RUN_PROC" | adcprep >>  "logs/runs/s${STORM_NUM}_pre.log"
+printf '%s\n2\n' "$ADCIRC_RUN_PROC" | ./adcprep >> $LOG_FILE
 log INFO "ADCPREP second execution done."
 
 log INFO "Finished pre-process for storm ${STORM_NUM}"
